Moved the move generation code out of page component
