+++ app/page.tsx
@@ -5,22 +5,8 @@ import React, { useCallback, useEffect, useReducer } from "react";
 import { Dashboard, Gameboard } from "./components";
 import { PieceId } from "./components/GamePiece";
 import { initialState } from "./data";
+import { findNextBestMove } from "./move-generator";
 import { reducer } from "./reducer";
-import { pieceWeights } from "./piece-weights";
-
-import {
-  calculateGamePieceMoves,
-  invertAndMapPieceState,
-} from "./move-options";
-
-/**
- * Describe a move for a given piece
- * 0 = piece to move
- * 1 = space to move to
- * 2 = piece that would be captured at that space (optional)
- * 3 = value of optional captured piece
- */
-type MoveSet = [PieceId, number, PieceId | null, number];
 
 export default function Home() {
   const [
@@ -56,6 +42,8 @@ export default function Home() {
     }
   }, [dispatch, inCheck, inCheckMate]);
 
+  // Once it's the black players turn, this effect will
+  // move their piece after a short timeout
   useEffect(() => {
     if (turn !== "black") {
       return;
@@ -64,87 +52,7 @@ export default function Home() {
     const randomWait = Math.ceil(Math.random() * 3);
 
     setTimeout(() => {
-      // Loop through all blk entries and generate paths
-      // Determine if there is competitor piece at any of them
-      // Capture a random piece or select a random move
-
-      const pieceMap = invertAndMapPieceState(activePieces);
-      const blackPieces = Object.keys(activePieces).filter((pieceId) =>
-        /blk/.test(pieceId)
-      );
-
-      const d: Array<MoveSet> = [];
-
-      for (const piece of blackPieces) {
-        const possibleMoves = calculateGamePieceMoves(
-          piece as PieceId,
-          activePieces
-        );
-        // If the piece can't move then go to the next piece
-        if (possibleMoves.length === 0) {
-          continue;
-        }
-
-        // [black piece, space to move to, piece it would capture, the captured value]
-
-        for (const move of possibleMoves) {
-          const possibleCapturePiece = pieceMap.get(move);
-          // If there is no piece, its worth nothing
-          if (possibleCapturePiece === undefined) {
-            d.push([piece as PieceId, move, null, 0]);
-          } else {
-            try {
-              const [_, pieceSlug] = /^wh-([a-z]+)\d{0,2}$/.exec(
-                possibleCapturePiece
-              )!;
-              const score =
-                pieceWeights[pieceSlug as keyof typeof pieceWeights];
-
-              d.push([piece as PieceId, move, possibleCapturePiece, score]);
-            } catch (e) {
-              console.log(
-                "This shouldnt happen, but going to catch incase the destruct fails",
-                e
-              );
-
-              d.push([piece as PieceId, move, null, 0]);
-            }
-          }
-        }
-      }
-
-      // Group these into an array where < index == more value
-      // Grab a random value from the 0 index and move there
-      // [
-      //   [
-      //     ['blk-b2', 40, 'wh-p1', 1]
-      //   ],
-      //   [
-      //     ['blk-p1', 24, null, 0],
-      //     ['blk-p3', 26, null, 0]
-      //   ]
-      // ]
-      // const movesSortedByWeight: Array<MoveSet[]> = []
-      const moveSetMap: Map<number, MoveSet[]> = new Map();
-      for (const moveSet of d) {
-        // Get the array of moves corresponding w/ this score
-        const scoreGroup = moveSetMap.get(moveSet[3]);
-
-        // If The score group doesn't exist, then create one
-        if (scoreGroup === undefined) {
-          moveSetMap.set(moveSet[3], [moveSet]);
-        } else {
-          moveSetMap.set(moveSet[3], [...scoreGroup, moveSet]);
-        }
-      }
-
-      const [bestMoves] = Array.from(moveSetMap.keys())
-        .sort()
-        .reverse()
-        .map((score) => moveSetMap.get(score) as Array<MoveSet>);
-
-      const randomIdx = Math.floor(Math.random() * bestMoves.length);
-      const [pieceToMove, nextPosition, pieceToCapture] = bestMoves[randomIdx];
+      const [pieceToMove, nextPosition, pieceToCapture] = findNextBestMove(activePieces)
 
       handlePiecePositionChange(pieceToMove, nextPosition, pieceToCapture ?? undefined)
     }, randomWait * 1000);
